+++ src/backend/libs/efiling-diligen-client/src/main/java/ca/bc/gov/open/efilingdiligenclient/diligen/DiligenAuthServiceImpl.java
@@ -1,0 +1,2 @@
+package ca.bc.gov.open.efilingdiligenclient.diligen;public class DiligenAuthServiceImpl {
+}
